NFC: Move FlattenTransferReadOp in preparation for reuse
This commit moves FlattenTransferReadOp to a common place in
preparation for reuse in the SPIR-V CodeGen pipeline.
