+++ compiler/src/iree/compiler/Codegen/Common/CMakeLists.txt
@@ -153,6 +153,7 @@ iree_cc_library(
     "DecomposeLinalgGeneric.cpp"
     "FoldAffineMinInDistributedLoops.cpp"
     "GPUDistributeSharedMemoryCopy.cpp"
+    "GPUPatterns.cpp"
     "GPUPipelining.cpp"
     "GPUTileReduction.cpp"
     "GPUVectorization.cpp"
